+++ docs/changelog.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+Added part to header to allow customization from outside the component.
+
 ### Changed
 
 ### Fixed