+++ docs/changelog.md
@@ -8,11 +8,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
-Added part to header to allow customization from outside the component.
+- Added part to header to allow customization from outside the component.
+- `invalid` and `error-text` properties added to `six-file-upload`
+- `active-tab-indicator` part for the indicator in the six-tab component
 
 ### Changed
 
+- Upgrade project to internally build with Node18
+
 ### Fixed
+
+- `six-error` examples not showing in the docs
+- Fire blur event when leaving multiselect component
+- Apply the debounce property from the date-picker to the time-picker
 
 ## 4.2.2 - 2024-03-28